+++ src/NeuralNet/CostFunctions/RelativeEntropy.php
@@ -1,0 +1,53 @@
+<?php
+
+namespace Rubix\ML\NeuralNet\CostFunctions;
+
+/**
+ * Relative Entropy
+ *
+ * Relative Entropy or *Kullback-Leibler divergence* is a measure of how the
+ * expectation and activation of the network diverge.
+ *
+ * @category    Machine Learning
+ * @package     Rubix/ML
+ * @author      Andrew DalPino
+ */
+class RelativeEntropy implements CostFunction
+{
+    /**
+     * Return a tuple of the min and max output value for this function.
+     *
+     * @return array
+     */
+    public function range() : array
+    {
+        return [-INF, INF];
+    }
+
+    /**
+     * Compute the cost.
+     *
+     * @param  float  $expected
+     * @param  float  $activation
+     * @return float
+     */
+    public function compute(float $expected, float $activation) : float
+    {
+        return $expected * log(($expected + self::EPSILON)
+            / ($activation + self::EPSILON));
+    }
+
+    /**
+     * Calculate the derivative of the cost function.
+     *
+     * @param  float  $expected
+     * @param  float  $activation
+     * @param  float  $computed
+     * @return float
+     */
+    public function differentiate(float $expected, float $activation, float $computed) : float
+    {
+        return ($activation - $expected + self::EPSILON)
+            / ($activation + self::EPSILON);
+    }
+}